Add nested aggregations inside connections in schema
