+++ packages/graphql/src/classes/Neo4jDatabaseInfo.ts
@@ -29,7 +29,7 @@ export class Neo4jDatabaseInfo {
     public edition: Neo4jEdition | undefined;
 
     constructor(version: string, edition?: Neo4jEdition) {
-        // Quick hack to support CalVar
+        // Quick hack to support CalVer
         version = version.replace(/\.0([0-9]+)/, ".$1");
         this.version = this.toSemVer(version);
         this.rawVersion = version;
